Refactor to move fancy output logic to realtime compiler class
